telegram/requirements.txt
python-telegram-bot==v20.0b0
qreader==1.3.1
python-dotenv==0.21.0